✨ Added Automatic Discovery of Video File Type:
- If a provided video file has an unknown file extension, the mod will attempt to find out what video type it is and create a temporary copy to use with the correct file extension.

+++ src/HelperFunctions/EmbedHelpers.cs
@@ -9,7 +9,7 @@ namespace NewSafetyHelp.HelperFunctions
 {
     public static class EmbedHelpers
     {
-        private const string NewSafetyHelpPrefix = "NewSafetyHelp_";
+        public const string NewSafetyHelpPrefix = "NewSafetyHelp_";
 
         /// <summary>
         /// Extracts the embedded resource to a temporary file.
@@ -91,8 +91,11 @@ namespace NewSafetyHelp.HelperFunctions
             {
                 try
                 {
-                    LoggingHelper.DebugLog($"Deleting temporary file '{tempFile}'.");
-                    File.Delete(tempFile);
+                    if (!string.IsNullOrEmpty(tempFile))
+                    {
+                        LoggingHelper.DebugLog($"Deleting temporary file '{tempFile}'.");
+                        File.Delete(tempFile);
+                    }
                 }
                 catch (Exception e)
                 {
